Fully tested and functioning code
Apologies in case nobody noticed the first commit was at 3 am the other day, I was not in my best of mind. ;)

I've since tested all of the code. I am including a screeny of said test success as proof. ;) Among other things, a few files got renamed and I've included ExampleObjecet.cs as, well as an instantiable example.